info

Provide a broad description of an object.

info(data, n = NULL, warn.print = TRUE)

Arguments

data

Object to analyse.

n

Single positive integer value indicating the n first number of elements to display per compartment of the output list (i.e., head(..., n)). Write NULL to return all the elements. Does not apply for the $STRUCTURE compartment output.

warn.print

Single logical value. Print potential warnings at the end of the execution? If FALSE the warning messages are added in the output list as an additional compartment (or NULL if no message).

Value

A list containing information, depending on the class and type of data. The backbone is generally: - $NAME: name of the object. - $CLASS: class of the object (class() value). - $TYPE: type of the object (typeof() value). - $LENGTH: length of the object (length() value). - $NA.NB: number of NA and NaN (only for type "logical", "integer", "double", "complex", "character" or "list"). - $HEAD: head of the object (head() value). - $TAIL: tail of the object (tail() value). - $DIMENSION: dimension (only for object with dimensions). - $SUMMARY: object summary (summary() value). - $STRUCTURE: object structure (str() value). - $WARNING: warning messages (only if the warn.print argument is FALSE).

If data is made of numerics, provide also: - $INF.NB: number of Inf and -Inf. - $RANGE: range after removing Inf and NA. - $SUM: sum after removing Inf and NA. - $MEAN: mean after removing Inf and NA.

If data is a 2D object, provide also:

- $ROW_NAMES: row names. - $COL_NAMES: column names.

If data is a data frame, provide also:

- $COLUMN_TYPE: type of each column (typeof() value).

If data is a list, provide also:

- $COMPARTMENT_NAMES: names of the comprtments. - $COMPARTMENT_TYPE: type of each compartment (typeof() value).

Details

REQUIRED PACKAGES

cuteDev

REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

arg_check()

get_message()

WARNINGS

None

Examples

info(data = 1:3)
#> $NAME
#> [1] "1:3"
#> 
#> $CLASS
#> [1] "integer"
#> 
#> $TYPE
#> [1] "integer"
#> 
#> $LENGTH
#> [1] 3
#> 
#> $INF.NB
#> [1] 0
#> 
#> $RANGE
#> [1] 1 3
#> 
#> $SUM
#> [1] 6
#> 
#> $MEAN
#> [1] 2
#> 
#> $NA.NB
#> [1] 0
#> 
#> $SUMMARY
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>     1.0     1.5     2.0     2.0     2.5     3.0 
#> 
#> $STRUCTURE
#>                  
#>   int [1:3] 1 2 3
#>