tail

As tail() but display the left or right head of big 2D objects.

tail(data1, n = 6, side = "l")

Arguments

data1

Any object but more dedicated for matrix, data frame or table.

n

As in tail() but for for matrix, data frame or table, number of dimension to print (10 means 10 rows and columns).

side

Either "l" or "r" for the left or right side of the 2D object (only for matrix, data frame or table).

Value

The tail.

Details

REQUIRED PACKAGES

cuteDev

REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

arg_check()

BEWARE

Other arguments of tail() not used.

Examples

obs1 = matrix(1:30, ncol = 5, dimnames = list(letters[1:6], LETTERS[1:5])) ; 
obs1 ; 
#>   A  B  C  D  E
#> a 1  7 13 19 25
#> b 2  8 14 20 26
#> c 3  9 15 21 27
#> d 4 10 16 22 28
#> e 5 11 17 23 29
#> f 6 12 18 24 30
tail(obs1, 3, "r")
#>    C  D  E
#> d 16 22 28
#> e 17 23 29
#> f 18 24 30